+++ apps/bare-expo/macos/BareExpo-macOS/AppDelegate.swift
@@ -4,15 +4,14 @@ import ReactAppDependencyProvider
 
 public class AppDelegate: ExpoAppDelegate {
   public override func applicationDidFinishLaunching(_ notification: Notification) {
-    self.moduleName = "main"
-    self.initialProps = [:]
-
     let delegate = ReactNativeDelegate()
     let factory = ExpoReactNativeFactory(delegate: delegate)
     delegate.dependencyProvider = RCTAppDependencyProvider()
 
     reactNativeFactoryDelegate = delegate
     reactNativeFactory = factory
+    
+    window = UIWindow(frame: UIScreen.main.bounds)
 
     return super.applicationDidFinishLaunching(notification)
   }